Generic functions in function decorators.
The function decorators need to handle the general case of functions and
are not supposed to be focused on continuous functions only.

+++ src/main/java/net/sourceforge/cilib/problem/FunctionMaximisationProblem.java
@@ -52,7 +52,7 @@ public class FunctionMaximisationProblem extends FunctionOptimisationProblem {
      */
     @Override
     protected Fitness calculateFitness(Type solution) {
-        return new MaximisationFitness(function.evaluate((Vector) solution));
+        return new MaximisationFitness(function.evaluate((Vector) solution).doubleValue());
     }
 
     /**
@@ -70,6 +70,6 @@ public class FunctionMaximisationProblem extends FunctionOptimisationProblem {
      */
     @Override
     public double getError(Type solution) {
-        return function.getMaximum() - function.evaluate((Vector) solution);
+        return function.getMaximum().doubleValue() - function.evaluate((Vector) solution).doubleValue();
     }
 }
